+++ axelor-business-support/src/main/java/com/axelor/apps/businesssupport/module/BusinessSupportModule.java
@@ -2,8 +2,10 @@ package com.axelor.apps.businesssupport.module;
 
 import com.axelor.app.AxelorModule;
 import com.axelor.apps.businessproject.db.repo.TeamTaskBusinessProjectRepository;
+import com.axelor.apps.businessproject.service.ProjectBusinessServiceImpl;
 import com.axelor.apps.businessproject.service.TeamTaskBusinessProjectServiceImpl;
 import com.axelor.apps.businesssupport.db.repo.TeamTaskBusinessSupportRepository;
+import com.axelor.apps.businesssupport.service.ProjectBusinessSupportServiceImpl;
 import com.axelor.apps.businesssupport.service.TeamTaskBusinessSupportServiceImpl;
 
 public class BusinessSupportModule extends AxelorModule {
@@ -12,5 +14,6 @@ public class BusinessSupportModule extends AxelorModule {
   protected void configure() {
     bind(TeamTaskBusinessProjectServiceImpl.class).to(TeamTaskBusinessSupportServiceImpl.class);
     bind(TeamTaskBusinessProjectRepository.class).to(TeamTaskBusinessSupportRepository.class);
+    bind(ProjectBusinessServiceImpl.class).to(ProjectBusinessSupportServiceImpl.class);
   }
 }